+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Changes
 
 ## UNRELEASED
 
-* Feature: Add `ignores` configuration option.
+* Feature: Add `ignores`, `allowMissing` configuration options.
   [#4](https://github.com/FormidableLabs/trace-pkg/issues/4)
 
 ## 0.1.0